add viewcore pprof command
This experimental command creates a pprof file that contains one
"sample" per path through the object graph, such that each path
terminates when it hits an object that's already contained in another
path.

To coax this data into pprof, each "file location" is simply the type of
the object in the path. This creates a pprof output that's essentially
the object graph grouped by type.